refactored commands, async functions not awaiting properly however.
