+++ index.js
@@ -1,12 +1,14 @@
-const axios = require('axios');
-const cctsAPI = axios.create({ baseURL: 'https://stormy-tundra-35633.herokuapp.com/' });
+require('dotenv').config();
 const Discord = require('discord.js');
 const client = new Discord.Client();
-var botResponse = new Discord.RichEmbed();
-const express = require('express')
-const app = express()
+// const express = require('express');
+// const app = express();
+const {commands} = require("./functions/commands");
+const {axiosPost} = require("./functions/axiosPost");
+const {axiosGet} = require("./functions/axiosGet");
+var user = {};
 
-app.listen(process.env.PORT)
+// app.listen(process.env.PORT);
 
 // Bot fails to opperate before this ready acknowledgement
 client.on('ready', () => {
@@ -19,146 +21,18 @@ client.on('message', m => {
     if (m.author.bot) { return }
 
     if (!m.content.startsWith(process.env.PREFIX)) {
-        cctsAPI.get(`/users/${m.author.id}`).then(response => {
-            if (response.data == []) {
-                cctsAPI.post('/users', { user_id: m.author.id, username: m.author.username, points: 1 })
-                    .then(res => {
-                        m.channel.send(`Congratulations ${res.data.username}! You have been noticed by ${client.user.username}.`)
-                    }).catch(function (error) {
-                        if (error.response) {
-                            // The request was made and the server responded with a status code
-                            console.log(error.response.data);
-                            console.log(error.response.status);
-                            console.log(error.response.headers);
-                        } else if (error.request) {
-                            // The request was made but no response was received
-                            console.log(error.request);
-                        } else {
-                            // Something happened in setting up the request that triggered an Error
-                            console.log('Error', error.message);
-                        }
-                        console.log(error.config);
-                    });
-            }
-        })
+        user = axiosGet(m.author.id)
+        if (!user.user_id) { 
+            user = axiosPost(m.author.id, m.author.username)
+            m.channel.send(`Congratulations ${res.data.username}! You have been noticed by ${client.user.username}.`)        
+        }   
         return
     }
     else {
-        // Seperating the user message into meaningful chunks 
-        mArray = m.content.split(" ")
-        mCommand = mArray[0].slice(1)
-        mArguments = mArray.slice(1)
-
-        // Process user message and respond accordingly
-        switch (mCommand) {
-            case 'botinfo':
-                botResponse = new Discord.RichEmbed()
-                    .setDescription('Bot Information')
-                    .setColor("#FFFFFF")
-                    .addField("Bot Name", client.user.username);
-                return m.channel.send(botResponse);
-
-            case 'register':
-                cctsAPI.post('/users', { user_id: m.author.id, username: m.author.username, points: 0 })
-                    .then(response => {
-                        m.channel.send(`Congratulations ${response.data.username}! You have successfully registered for the community points program.`)
-                    })
-                    .catch(function (error) {
-                        if (error.response) {
-                            // The request was made and the server responded with a status code
-                            console.log(error.response.data);
-                            console.log(error.response.status);
-                            console.log(error.response.headers);
-                        } else if (error.request) {
-                            // The request was made but no response was received
-                            console.log(error.request);
-                        } else {
-                            // Something happened in setting up the request that triggered an Error
-                            console.log('Error', error.message);
-                        }
-                        console.log(error.config);
-                    });
-                return;
-
-            case 'points':
-                cctsAPI.get(`/users/${m.author.id}`).then(response => {
-                    m.channel.send(`${m.author.username}, you have ${response.data.points} points!`)
-                }).catch(function (error) {
-                    if (error.response) {
-                        // The request was made and the server responded with a status code
-                        console.log(error.response.data);
-                        console.log(error.response.status);
-                        console.log(error.response.headers);
-                    } else if (error.request) {
-                        // The request was made but no response was received
-                        console.log(error.request);
-                    } else {
-                        // Something happened in setting up the request that triggered an Error
-                        console.log('Error', error.message);
-                    }
-                    console.log(error.config);
-                });
-                return;
-
-            case 'update':
-                cctsAPI.get(`/users/${m.author.id}`).then(response => {
-                    if (response.data.username == m.author.username) {
-                        m.channel.send('Username already up to date! Nice!')
-                    }
-                    else {
-                        cctsAPI.patch(`/users/${m.author.id}`, { username: m.author.username }).then(res => {
-                            return m.channel.send(`${response.data.username} shall now be known as ${res.data.username}`)
-                        }).catch(function (error) {
-                            if (error.response) {
-                                // The request was made and the server responded with a status code
-                                console.log(error.response.data);
-                                console.log(error.response.status);
-                                console.log(error.response.headers);
-                            } else if (error.request) {
-                                // The request was made but no response was received
-                                console.log(error.request);
-                            } else {
-                                // Something happened in setting up the request that triggered an Error
-                                console.log('Error', error.message);
-                            }
-                            console.log(error.config);
-                        });
-                    }
-                }).catch(function (error) {
-                    if (error.response) {
-                        // The request was made and the server responded with a status code
-                        console.log(error.response.data);
-                        console.log(error.response.status);
-                        console.log(error.response.headers);
-                    } else if (error.request) {
-                        // The request was made but no response was received
-                        console.log(error.request);
-                    } else {
-                        // Something happened in setting up the request that triggered an Error
-                        console.log('Error', error.message);
-                    }
-                    console.log(error.config);
-                });
-                return;
-
-            case 'help':
-                botResponse = new Discord.RichEmbed()
-                    .setDescription('Valid Commands')
-                    .setColor("#FFFFFF")
-                    .addField(`${process.env.PREFIX}botinfo`, "Displays detailed information about the bot in use.")
-                    .addField(`${process.env.PREFIX}points`, 'Displays the number of community contribution points acumulated.')
-                    .addField(`${process.env.PREFIX}register`, 'Registers the user for community points program.')
-                    .addField(`${process.env.PREFIX}update`, 'Updates the users name in the CCTS. Please use this if you have recently changed your name recently as the bot has a good memory and bad people skills.')
-                    .addField(`${process.env.PREFIX}help`, 'Displays this helpful little list of commands for the uninitiated.');
-
-                return m.channel.send(botResponse);
-
-            default:
-                return m.channel.send(`Not a valid bot request, type ${process.env.PREFIX}help for a list of valid commands.`)
-        }
+        // User input is a command and therefore must be handled by the commands function
+        commands(m);
     }
 })
-
 
 
 client.login(process.env.TOKEN);
